Graphics Engine Rework Checkpoint
Rework the style class. Each SceneNode now has a mutable Style instance. The style can be updated in real time. This is needed for the future
animation engine and supporting manually changing style in user code. The Style also can function as a "static" style class. The user creates style sheets,
which are collections of immutable style class instances. The static instances can be assigned to SceneNode styles (this is for react style property support).

Moved ImageStoreView, FontStoreView and Logger into a bindings package (code organization).