Patch GC Crash
ObjectWrap constructor creates a js object in the vm. The object has a finalizer that deletes the ObjectWrap this ptr when the js object is
eligible for gc. If the subclass (or ObjectWrap) constructor throws an exception, the this ptr is invalidated by C++. However, the pointer
is still associated with the finalizer registered with the vm. No way to unregister the finalizer.

For now, the limited light source javascript objects have been patched to reduce the chances of them throwing exceptions.. so the tests
can run in CI. The longer term solution is to address the problem in ObjectWrap or build a custom ObjectWrap.