Fix scheduler interoperability issue
This is an ugly hack to get scheduler working. For the future, it would be best to create a custom scheduler for light
source (instance-able, clear lifecycle and knows about scene/stage paradigm).

scheduler depends on global window and MessageChannel for initialization. The necessary globals are polyfilled
in. Though, exposing a partially populated window object may create problems for other libraries. MessageChannel
has open ports that the scheduler does not shutdown. A pseudo-singleton class was created to address closing the
ports.