Restructure Plugin System
The plugins for audio and platform used to be separate addons. light-source.node would interact with these addons and static libraries were shared
with all addons. This was difficult, if not impossible, to manage at compile time and run time across gcc, clang and msvc. After some research on the
C++ abi and shared libraries, the design decision of multiple addons (shared libraries) is the issue. The solution is to preserve the plugin system, but
move all of the code to one addon (shared library), light-source.node. The plugin interfaces have more or less remained unchanged. sdl, sdl-audio,
sdl-mixer and ref can all be dynamically loaded in javascript.

As part of this change, I manually load SDL and SDL_mixer using dlopen and get the functions I need. With this loading strategy, I can load light-source.node
and deal with missing SDL libraries in the javascript runtime.