Support new JSX transform
react/jsx-transform is in cjs format and require()'s react, resulting in two copies of react in memory when other parts of the
app imports react. To work around this problem, @lse/react will export an ESM version of jsx-transform. transpilation
tools will need to use @lse/react/jsx-transform. If react exports an ESM version, the rollup patch of jsx-transform can
be removed.

Updated lse-node and examples to use the new JSX transform.