Fix input manager bugs
* hat events were not always getting mapped
* axis motion events and axis mapping were not implemented
* keyboard mapping did not work with mapped analog keys
* uncovered other misc bugs in the input manager with more unit testing

The underlying issue with most of these bugs is how mapping of input device
events are handled. The original idea was to implement the mapping in javascript
to have flexibility to the mapping api. The flexibility was not necessary, so the mapping
is now done by the SDL game controller api. This change had a ripple effect on the
input manager and it's public apis. The biggest change is mappings. The implementation
now treats the installed game controller mappings as "intrinsic". Through the input manager,
user defined mappings can be override the intrinsic mappings. I think the mappings now
better address the use case of the user customizing gamepads (mappings, serialize and
deserialize).

Some issues remain: axis events, gamecontrollerdb loading